Replace file.
