+++ src/lib/win32lib.py
@@ -1,4 +1,4 @@
-# Provides bindings to lib.dll
+# Provides bindings to [lib.dll].
 
 from ctypes import CDLL, c_char_p
 from os.path import dirname
@@ -7,21 +7,17 @@ from os.path import dirname
 dll = CDLL(f"{dirname(__file__)}/lib.dll")
 
 
-def filedialog(title: str = "Open", filters: str = "", dir: str = "") -> list[str]:
+def openFileDialog(title: str = "Open", filters: str = "All Files (*.*)|*.*", dir: str = ".") -> list[str]:
     # Title -> Title of the file dialog.
     # Filters -> Files (*.ext1, *.ext2, ...)|*.ext1;*.ext2;...
     # Directory -> Set the directory to open the file dialog in.
 
-    fd = dll.filedialog
-    fd.restype = c_char_p
-    fd.argtypes = [c_char_p, c_char_p, c_char_p]
-    return fd(title.encode(), filters.encode(), dir.encode()).decode('UTF-8').splitlines()
+    ofd = dll.openFileDialog
+    ofd.restype = c_char_p
+    ofd.argtypes = [c_char_p, c_char_p, c_char_p]
+    return ofd(title.encode(), filters.encode(), dir.encode()).decode().splitlines()
 
 
-def set_sm_on_top():
-    # Puts the Smoothie Window on top of all other windows.
-    dll.setSMOnTop()
-
-def set_sm_debug():
-    # Puts the Smoothie Window on top of all other windows.
-    dll.setSMDebug()
+def setWinOnTop():
+    # Puts any foreground window on top of all other windows.
+    dll.setWinOnTop()
